# api/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateExperimentRunRequest": {
        "properties": {
          "candidate_version": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Candidate Version",
            "type": "string"
          },
          "eval_case_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eval Case Ids"
          },
          "eval_spec_id": {
            "format": "uuid",
            "title": "Eval Spec Id",
            "type": "string"
          },
          "tenant_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          }
        },
        "required": [
          "eval_spec_id",
          "candidate_version"
        ],
        "title": "CreateExperimentRunRequest",
        "type": "object"
      },
...
      "ExperimentRun": {
        "properties": {
          "candidate_version": {
            "title": "Candidate Version",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "eval_spec_id": {
            "format": "uuid",
            "title": "Eval Spec Id",
            "type": "string"
          },
          "experiment_run_id": {
            "format": "uuid",
            "title": "Experiment Run Id",
            "type": "string"
          },
          "ingest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExperimentRunIngest"
              },
              {
                "type": "null"
              }
            ]
          },
          "result_count": {
            "title": "Result Count",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ExperimentRunStatus"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "experiment_run_id",
          "tenant_id",
          "eval_spec_id",
          "candidate_version",
          "status",
          "result_count",
          "created_at",
          "updated_at"
        ],
        "title": "ExperimentRun",
        "type": "object"
      },
      "ExperimentRunIngest": {
        "properties": {
          "artifact_paths": {
            "additionalProperties": true,
            "title": "Artifact Paths",
            "type": "object"
          },
          "eval_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eval Summary"
          },
          "external_run_id": {
            "title": "External Run Id",
            "type": "string"
          },
          "failure_packet": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Packet"
          },
          "gate_explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Explanation"
          },
          "gate_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Status"
          },
          "outputs": {
            "additionalProperties": true,
            "title": "Outputs",
            "type": "object"
          },
          "scenario_ids": {
            "items": {
              "type": "string"
            },
            "title": "Scenario Ids",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "suite_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suite Id"
          }
        },
        "required": [
          "source",
          "external_run_id"
        ],
        "title": "ExperimentRunIngest",
        "type": "object"
      },
      "ExperimentRunListResponse": {
        "properties": {
          "experiment_runs": {
            "items": {
              "$ref": "#/components/schemas/ExperimentRun"
            },
            "title": "Experiment Runs",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "experiment_runs",
          "request_id"
        ],
        "title": "ExperimentRunListResponse",
        "type": "object"
      },
      "ExperimentRunResponse": {
        "properties": {
          "candidate_version": {
            "title": "Candidate Version",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "eval_spec_id": {
            "format": "uuid",
            "title": "Eval Spec Id",
            "type": "string"
          },
          "experiment_run_id": {
            "format": "uuid",
            "title": "Experiment Run Id",
            "type": "string"
          },
          "ingest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExperimentRunIngest"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "result_count": {
            "title": "Result Count",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ExperimentRunStatus"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "experiment_run_id",
          "tenant_id",
          "eval_spec_id",
          "candidate_version",
          "status",
          "result_count",
          "created_at",
          "updated_at",
          "request_id"
        ],
        "title": "ExperimentRunResponse",
        "type": "object"
      },
      "ExperimentRunStatus": {
        "enum": [
          "completed",
          "failed"
        ],
        "title": "ExperimentRunStatus",
        "type": "string"
      },
      "ExperimentRunSummaryResponse": {
        "properties": {
          "average_score": {
            "title": "Average Score",
            "type": "number"
          },
          "candidate_version": {
            "title": "Candidate Version",
            "type": "string"
          },
          "eval_spec_id": {
            "format": "uuid",
            "title": "Eval Spec Id",
            "type": "string"
          },
          "experiment_run_id": {
            "format": "uuid",
            "title": "Experiment Run Id",
            "type": "string"
          },
          "failed_count": {
            "title": "Failed Count",
            "type": "integer"
          },
          "pass_rate": {
            "title": "Pass Rate",
            "type": "number"
          },
          "passed_count": {
            "title": "Passed Count",
            "type": "integer"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "result_count": {
            "title": "Result Count",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ExperimentRunStatus"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "experiment_run_id",
          "tenant_id",
          "eval_spec_id",
          "candidate_version",
          "status",
          "result_count",
          "passed_count",
          "failed_count",
          "pass_rate",
          "average_score",
          "request_id"
        ],
        "title": "ExperimentRunSummaryResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QualityGateResponse": {
        "properties": {
          "average_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Score"
          },
          "candidate_version": {
            "title": "Candidate Version",
            "type": "string"
          },
          "eval_spec_id": {
            "format": "uuid",
            "title": "Eval Spec Id",
            "type": "string"
          },
          "evaluation_source": {
            "title": "Evaluation Source",
            "type": "string"
          },
          "experiment_run_id": {
            "format": "uuid",
            "title": "Experiment Run Id",
            "type": "string"
          },
          "external_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Run Id"
          },
          "gate_explanation": {
            "title": "Gate Explanation",
            "type": "string"
          },
          "gate_status": {
            "title": "Gate Status",
            "type": "string"
          },
          "ingest_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingest Source"
          },
          "pass_threshold": {
            "title": "Pass Threshold",
            "type": "number"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "experiment_run_id",
          "eval_spec_id",
          "candidate_version",
          "gate_status",
          "gate_explanation",
          "evaluation_source",
          "pass_threshold",
          "request_id"
        ],
        "title": "QualityGateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Control plane for eval-driven AI development on top of Langfuse.",
    "title": "Eval Driven Design Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/experiment-runs": {
      "get": {
        "operationId": "list_experiment_runs_v1_experiment_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "eval_spec_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Eval Spec Id"
            }
          },
          {
            "in": "query",
            "name": "ingest_source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ingest Source"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Experiment Runs",
        "tags": [
          "experiment-runs"
        ]
      },
      "post": {
        "operationId": "create_experiment_run_v1_experiment_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperimentRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Experiment Run",
        "tags": [
          "experiment-runs"
        ]
      }
    },
    "/v1/experiment-runs/{experiment_run_id}": {
      "get": {
        "operationId": "get_experiment_run_v1_experiment_runs__experiment_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Experiment Run",
        "tags": [
          "experiment-runs"
        ]
      }
    },
    "/v1/experiment-runs/{experiment_run_id}/gate": {
      "get": {
        "operationId": "get_experiment_run_gate_v1_experiment_runs__experiment_run_id__gate_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityGateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Experiment Run Gate",
        "tags": [
          "experiment-runs"
        ]
      }
    },
    "/v1/experiment-runs/{experiment_run_id}/summary": {
      "get": {
        "operationId": "get_experiment_run_summary_v1_experiment_runs__experiment_run_id__summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentRunSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Experiment Run Summary",
        "tags": [
          "experiment-runs"
        ]
      }
    },
...
  }
}